+++ WebLearning.Application/ELearning/Services/QuizMonthlyService.cs
@@ -166,8 +166,8 @@ namespace WebLearning.Application.ELearning.Services
                 pageCount = Math.Ceiling(query.Count() / (double)pageResult);
             }
             var totalRow = await query.CountAsync();
-            var data = await query.Skip((getListPagingRequest.PageIndex - 1) * (int)pageResult)
-                                    .Take((int)pageResult)
+            var data = await query.Skip((getListPagingRequest.PageIndex - 1) * pageResult)
+                                    .Take(pageResult)
                                     .Select(quizMonthly => new QuizMonthlyDto()
                                     {
                                         ID = quizMonthly.ID,
